+++ src/shared/ipc/commands/mod.rs
@@ -3,6 +3,7 @@ use std::{path::PathBuf, time::Duration};
 use anyhow::{Context, Result};
 use index_lrc::IndexLrcCommand;
 use keybind::KeybindCommand;
+use query_tab::QueryCommand;
 use set::SetIpcCommand;
 use status_message::StatusMessageCommand;
 use switch_tab::SwitchTabCommand;
@@ -16,22 +17,23 @@ use crate::config::{
 
 pub(super) mod index_lrc;
 pub mod keybind;
+pub(super) mod query_tab;
 pub(super) mod set;
 pub(super) mod status_message;
 pub mod switch_tab;
 pub(super) mod tmux;
 
-impl TryFrom<RemoteCmd> for SocketCommand {
+impl TryFrom<&RemoteCmd> for SocketCommand {
     type Error = anyhow::Error;
 
-    fn try_from(value: RemoteCmd) -> Result<Self> {
+    fn try_from(value: &RemoteCmd) -> Result<Self> {
         log::debug!(value:?; "Got remote command");
 
         match value {
-            RemoteCmd::IndexLrc { ref path } => {
+            RemoteCmd::IndexLrc { path } => {
                 Ok(SocketCommand::IndexLrc(IndexLrcCommand { path: path.clone() }))
             }
-            RemoteCmd::Status { ref message, level, timeout } => {
+            RemoteCmd::Status { message, level, timeout } => {
                 Ok(SocketCommand::StatusMessage(StatusMessageCommand {
                     level: match level {
                         crate::config::cli::Level::Info => crate::shared::events::Level::Info,
@@ -39,10 +41,12 @@ impl TryFrom<RemoteCmd> for SocketCommand {
                         crate::config::cli::Level::Warn => crate::shared::events::Level::Warn,
                     },
                     message: message.clone(),
-                    timeout: Duration::from_millis(timeout),
+                    timeout: Duration::from_millis(*timeout),
                 }))
             }
-            RemoteCmd::Tmux { hook } => Ok(SocketCommand::TmuxHook(TmuxHookCommand { hook })),
+            RemoteCmd::Tmux { hook } => {
+                Ok(SocketCommand::TmuxHook(TmuxHookCommand { hook: hook.clone() }))
+            }
             RemoteCmd::Set { command: SetCommand::Config { path } } if path == "-" => {
                 Ok(SocketCommand::Set(Box::new(SetIpcCommand::Config(ron::de::from_reader(
                     std::io::stdin(),
@@ -66,8 +70,15 @@ impl TryFrom<RemoteCmd> for SocketCommand {
 
                 Ok(SocketCommand::Set(Box::new(SetIpcCommand::Theme(ron::de::from_reader(read)?))))
             }
-            RemoteCmd::Keybind { key } => Ok(SocketCommand::Keybind(KeybindCommand { key })),
-            RemoteCmd::SwitchTab { tab } => Ok(SocketCommand::SwitchTab(SwitchTabCommand { tab })),
+            RemoteCmd::Keybind { key } => {
+                Ok(SocketCommand::Keybind(KeybindCommand { key: key.clone() }))
+            }
+            RemoteCmd::SwitchTab { tab } => {
+                Ok(SocketCommand::SwitchTab(SwitchTabCommand { tab: tab.clone() }))
+            }
+            RemoteCmd::Query { targets } => {
+                Ok(SocketCommand::Query(QueryCommand { targets: targets.clone() }))
+            }
         }
     }
 }