+++ src/shared/ipc/commands/query_tab.rs
@@ -5,24 +5,26 @@ use serde::{Deserialize, Serialize};
 use crate::{
     AppEvent,
     WorkRequest,
-    config::Config,
+    config::{Config, cli::RemoteCommandQuery},
     shared::ipc::{IpcStream, SocketCommandExecute},
 };
 
 #[derive(Debug, Deserialize, Serialize)]
-pub(crate) struct TmuxHookCommand {
-    pub(crate) hook: String,
+pub(crate) struct QueryCommand {
+    pub targets: Vec<RemoteCommandQuery>,
 }
 
-impl SocketCommandExecute for TmuxHookCommand {
+impl SocketCommandExecute for QueryCommand {
     fn execute(
         self,
         event_tx: &Sender<AppEvent>,
         _work_tx: &Sender<WorkRequest>,
-        _stream: IpcStream,
+        stream: IpcStream,
         _config: &Config,
     ) -> Result<()> {
-        event_tx.send(AppEvent::TmuxHook { hook: self.hook })?;
+        event_tx
+            .send(AppEvent::IpcQuery { stream, targets: self.targets })
+            .map_err(|err| anyhow::anyhow!("Failed to send QueryTab event: {}", err))?;
         Ok(())
     }
 }