+++ src/shared/mod.rs
@@ -1,6 +1,7 @@
 pub mod dependencies;
 pub mod env;
 pub mod events;
+pub mod exit_code;
 pub mod ext;
 pub mod geometry;
 pub mod id;